requirements.txt
yt-dlp
google-genai
fastapi
uvicorn
pydantic-settings
tenacity
httpx
